refactor: restructure Home component and add webflow image
